plantcare-ai/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
tensorflow
numpy
pillow
python-multipart
